+++ edx_proctoring/tests/test_views.py
@@ -1836,8 +1836,9 @@ class TestStudentProctoredExamAttempt(LoggedInTestCase):
 
         # create number of exam attempts
         for i in range(90):
+            user = User.objects.create(username='student{0}'.format(i), email='student{0}@test.com'.format(i))
             ProctoredExamStudentAttempt.create_exam_attempt(
-                proctored_exam.id, i, 'test_name{0}'.format(i),
+                proctored_exam.id, user.id, 'test_name{0}'.format(i),
                 'test_attempt_code{0}'.format(i), True, False, 'test_external_id{0}'.format(i)
             )